FINALLY! add capability to correctly Grade a homework(Square.java is tested!)
In order to make TestCase Square.java work, I had to delete
System.outs because of parsing the output of the homework. We may
require some other configuration for the homework. (for parsing the
output) I cleaned some of the comments and files.  Finally I made
Config.java to be abstract, we should consider making Homework.java
also abstract since without inheriting it is useless.